upcoming: [DI-20933] - Moved common methods to utils file
